Move specific classes for FDroid into FDroid variant source code
