+++ vector/src/fdroid/java/im/vector/riotredesign/fdroid/features/settings/troubleshoot/TestBackgroundRestrictions.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.riotredesign.push.fcm.troubleshoot
+package im.vector.riotredesign.fdroid.features.settings.troubleshoot
 
 import android.content.Context
 import android.net.ConnectivityManager
