+++ vector/src/main/java/im/vector/riotredesign/VectorApplication.kt
@@ -36,7 +36,6 @@ import com.github.piasy.biv.loader.glide.GlideImageLoader
 import com.jakewharton.threetenabp.AndroidThreeTen
 import im.vector.matrix.android.api.Matrix
 import im.vector.riotredesign.core.di.AppModule
-import im.vector.riotredesign.core.services.AlarmSyncBroadcastReceiver
 import im.vector.riotredesign.features.configuration.VectorConfiguration
 import im.vector.riotredesign.features.crypto.keysbackup.KeysBackupModule
 import im.vector.riotredesign.features.home.HomeModule
@@ -47,7 +46,6 @@ import im.vector.riotredesign.features.notifications.PushRuleTriggerListener
 import im.vector.riotredesign.features.rageshake.VectorFileLogger
 import im.vector.riotredesign.features.rageshake.VectorUncaughtExceptionHandler
 import im.vector.riotredesign.features.roomdirectory.RoomDirectoryModule
-import im.vector.riotredesign.features.settings.PreferencesManager
 import im.vector.riotredesign.features.version.getVersion
 import im.vector.riotredesign.push.fcm.FcmHelper
 import org.koin.android.ext.android.get
@@ -117,7 +115,7 @@ class VectorApplication : Application() {
 
             @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
             fun entersForeground() {
-                AlarmSyncBroadcastReceiver.cancelAlarm(appContext)
+                FcmHelper.onEnterForeground(appContext)
                 Matrix.getInstance().currentSession?.also {
                     it.stopAnyBackgroundSync()
                 }
@@ -129,20 +127,8 @@ class VectorApplication : Application() {
 
                 notificationDrawerManager.persistInfo()
 
-                if (FcmHelper.isPushSupported()) {
-                    //TODO FCM fallback
-                } else {
-                    //TODO check if notifications are enabled for this device
-                    //We need to use alarm in this mode
-                    if (PreferencesManager.areNotificationEnabledForDevice(applicationContext)) {
-                        if (Matrix.getInstance().currentSession != null) {
-                            AlarmSyncBroadcastReceiver.scheduleAlarm(applicationContext, 4_000L)
-                            Timber.i("Alarm scheduled to restart service")
-                        }
-                    }
-                }
+                FcmHelper.onEnterBackground(appContext, Matrix.getInstance().currentSession != null)
             }
-
         })
